cleanup(CDCL): Get rid of `repush` list
When cancelling decisions, we collect literals at a higher decision
level than the one we are cancelling to (when the `--minimal-bj` option
is enabled) into a list, then replay them at the new decision level
(using `enqueue_assigned`).

This method introduces linear space overhead for the intermediate list
that gets created to store the literals to repush, aptly named `repush`.

This patch uses a more efficient implementation with no additional
memory usage by directly moving the literal towards the start of the
vector in a single pass, and removes the `enqueue_assigned` function and
`repush` list.