CP: Support for domains on non-leaves
This patch re-adds support for storing domains associated with arbitrary
semantic values rather than only with leaves. In a way, it partially
reverts #1004 because some domains are not inherently structural and
cannot be stored on leaves only (for instance, we want to support
interval domains for bit-vectors, but constraints on a concatenation
cannot be propagated exactly on the elements of the concatenation).

Domains at leaves and non-leaves are connected through _structural
propagation_: when the domain of a leaf changes, the change is
propagated to the domain of the non-leaves that contains it using the
`map_leaves` function; when the domain of a non-leaf changes, the change
is propagated to the domain of the leaves it contains using the
`fold_leaves` function.