Add `next_split` and save splits in `better_bj`
The `next_decisions` mechanism was intended to be shared among case
split and a potential future move of Alt-Ergo towards mcSAT strategies.
However, due to the way case splits are currently processed, using a
list is somewhat of an issue: each propagation during the initial
processing phase (before `unsat` is called) will cause the same split to
be added to the list.

Hence, we introduce a separate `next_split` mechanism for case splits
that only remembers the last split.

Also, we need to save the splits/decisions/objectives in
`try_to_backjump_further`, otherwise we could end up making decisions on
values that are no longer in the theory.