Use a wrapper to set timers
We wrap some important functions with timers in order to profile their
calls but we don't use a dedicated function to do this job everywhere in
the codebase. In particular, there is some places where we don't ensure
the timer is correctly paused even if the wrapped function raises an
exception.

This PR adds a new wrapper `with_timer` to perform this job correctly.

Add timers for the ADT and Bitv theories too.