chore: Check modified files for style in pre-commit hook (#1042)
The pre-commit hook no longer calls the check_indentation.sh and
check_style.sh because those check *all* the files in the repository,
which makes commit operations slow. Instead, the pre-commit hook simply
calls ocp-indent on the  changed files, making for a smoother
experience.

This patch gives the style checks the same treatment, and calls the
ocp-checker binary only on the changed files in the pre-commit hook.
Since the code for the ocp-checker binary is in the repo, the commit
hook checks that it has been built and supports diffing from standard
input (in order to ensure that we actually test what will be committed
and not what is currently in the working directory); otherwise, it emits
a warning message and does not check the style.