feat: Statically linked binary releases
This patch modifies the Docker workflow to build a statically linked
portable binary of Alt-Ergo, following [1]. The workflow also now builds
statically linked macOS binaries for x86-64 and arm64 platforms.

Building the Linux binaries is fairly quick and hence performed on
push/pull_request. On the other hand, the macOS build is fairly slow,
and only performed when manually requested.

[1] : https://ocamlpro.com/blog/2021_09_02_generating_static_and_portable_executables_with_ocaml/