fix(dolmen): Preserve trigger order
In spite of what the comment says, the Dolmen frontend currently
reverses the order in which triggers are considered. This patch ensures
that the trigger order is preserved by using `List.map` instead of
`List.rev_map` (there is usually only very few triggers and the
`mk_expr` itself is not tail-recursive).