feat(BV,CP): Use a FIFO queue for constraint propagation
This patch changes the bit-vector propagator to use a FIFO queue to
select which propagator to run. Using a FIFO queue gives fairness
guarantees on the constraints that get propagated, and does not depend
on hash-consing shenanigans as in the current implementation.

The implementation is a simple wrapper around the `Queue` module from
the stdlib and a hash table to ensure uniqueness of constraints in the
queue.