feat(BV,CP): Ephemeral API for domain updates
This patch adds a new ephemeral (imperative) interface to the `Domains`
API, and uses it for constraint propagation. The conversion between the
persistent and ephemeral APIs is provided by a function `edit` that
returns an ephemeral domain (implemented using a hash table) from a
persistent one, and a `snapshot` function that performs the inverse
operation.

Ephemeral domains are only used internally during constraint propagation
so that the external API of the `Bitv_rel` module stays persistent.

This leads to more readable code for the constraints, and should be more
efficient on large problems as we do not keep looking up the domains in
a `Map`.