feat(CDCL): Add ability to decide on semantic literals (#1027)
This patch introduces the ability for the CDCL solver to deal with
semantic literals. This is part of work towards #901 although it does
not (yet) achieves the goals of #901.

More precisely:

 - We add a new `Literal` module in `structures` to represent literals
   that are either syntaxic (represented by expressions) or semantic
   (represented using the `Xliteral` module).

 - The new `Literal` module is used to represent the literals associated
   with the `Atom`s in the CDCL solver [1]

 - The CDCL solver has a special handling of these semantic literals,
   which are treated as "local" literals: they are not added in the
   variable ordering, and will not be decided upon by the solver.
   Instead, a new mechanism inspired by mSAT (and, transitively, mcSAT)
   is provided, where the theory is able to mark certain literals as
   "immediate decisions". When literals are marked this way, the solver
   will decide on one of the marked literals the next time a decision
   must be made. Literals marked as immediate decisions are part of a
   transient solver state, and are forgotten upon backtracking.

 - The theory uses this marking mechanism to communicate case splits
   (and a similar mechanism for optimization objectives, which is
   separate because the value of the objective is sent back to the
   theory once the decision is made) to the solver.

This mechanism allows to introduce splits that are, in a way, "local",
and do not  pollute other branches of the solver. It can be seen as an
intermediate step towards a mcSAT approach, but where we are still
deciding on boolean literals and hence try to avoid adding a large
number of splits globally into the solver -- although it does mean that
the splits are not incorporated into the activity heuristic.

The functionality is currently gated behind a new flag
`--enable-sat-cs`.

[1] : It could also be used for the `Bj`s in the Tableaux solver, if we
      wanted to suport similar features in the Tableaux solver.

* Add `next_split` and save splits in `better_bj`

The `next_decisions` mechanism was intended to be shared among case
split and a potential future move of Alt-Ergo towards mcSAT strategies.
However, due to the way case splits are currently processed, using a
list is somewhat of an issue: each propagation during the initial
processing phase (before `unsat` is called) will cause the same split to
be added to the list.

Hence, we introduce a separate `next_split` mechanism for case splits
that only remembers the last split.

Also, we need to save the splits/decisions/objectives in
`try_to_backjump_further`, otherwise we could end up making decisions on
values that are no longer in the theory.