feat(plugins): Use dune-site for inequalities plugins (#1049)
We already use dune-site to manage plugin files. With this patch, we now
also use the dune-site mechanism to manage plugin loading for the
inequalities plugin. This should make alt-ergo work with bytecode-only
compilers again (see [1]).

We still use Dynlink directly for additional parsers. Support for parser
plugins is deprecated as it is not supported with the Dolmen frontend
and will be removed in a future release, so the churn doesn't seem worth
it.

Fixes #1047

[1] : https://github.com/ocaml/opam-repository/pull/25294