fix(Dolmen): Parse triggers from formula body (#1051)
Triggers are annotated on the body of the quantifier, but Alt-Ergo looks
them up on the quantifier itself. This used to be OK, because Dolmen
used to copy tags from the body to the quantifier.

Since https://github.com/Gbury/dolmen/pull/207 Dolmen no longer performs
this copy, which means that we now ignore triggers (or rather, we use
the triggers of the parent quantifier if it exists -- which is wrong).

This patch looks for trigger on the body instead, to be compatible with
future Dolmen releases.

This patch also bumps the Dolmen version for obvious reasons.