fix(AC): Abstract AC arguments that contain other AC symbols (#990)
This patch implements a variation of the abstraction mechanism described
in the AC(X) paper. This should hopefully ensure that we can't create
substitution cycles due to improper term ordering.

Note that the issue is related to comparing distinct AC symbols: when
the AC symbols are identical, the AC theory uses a multiset ordering on
its argument, which prevent substitution cycles. But when the AC symbols
are different, only the symbols are compared, not the arguments, and so
we must rely on the abstraction mechanism to prevent cycles.

This is a do-over of https://github.com/OCamlPro/alt-ergo/pull/974 that
should be closer in spirit to the implementation of the paper and
without the associated regressions.

Fixes #964