fix(CP): Do not add unnecessary explanations to domains
When performing a substitution that results in rewriting [r] into [nr]
with explanation [ex], and [nr] does not yet have a registered domain,
explanations from [ex] and the domain of [r] should only be added if
they actually shrink the [nr]'s default domain.

Otherwise, we can end up with a constant having a non-empty explanation
for why it is equal to itself -- which is wholly unnecessary and can
cause irrelevant backtracking.