Sort trigger patterns before caching
As we noticed a long time ago, the `Matching` module of Alt-Ergo sorts
patterns of multi-triggers during each iteration of the matching round.
It's inefficient as the weight used for sorting doesn't change.

I thought we should perform these sorts during the generation of
triggers in `Expr` but actually, Alt-Ergo can generate new triggers
later depending on the matching configuration (see `Util.matching_env`).

In this PR, we perform the sorts just before caching the generated
triggers in `Matching`.