Preserve default domain during updates
The previous implementation of `Domain.update` could replace a default
domain (with the empty explanation) by a default one with a non-empty
explanation. There is no reason to add these extra explanations.

This new implementation doesn't update the map `t.domains` if the
intersection is a default domain. It means `update` doesn't add a domain
if `d` is a default domain and `r` isn't in the map, which is ok as we
ensure that all seen semantic values are added to `t.domains` in `assume`.