Fully disable heavy assertions
In some part of the codebase, we use heavy assertions as follow:
```
assert (not @@ Options.get_enable_assertions () || heavy_assertion)
```
This trick has two advantages:
- As the logical `or` in OCaml has short-circuit evaluation, if the option
  `enable_assertions` is not `true`, we don't evaluate the heavy
  assertion.
- If we use the compiler option `no-asserts`, we don't even test the
  option `Options.get_enable_assertions`.

This PR adds a new function in `Options` to reproduce this good design
everywhere in the codebase.

Notice that I put the function `heavy_assert` in `Options` but it would
be better to put it in `Util`. Alas, we got a cycle dependency :/