feat(CDCL-Tableaux): Do not make irrelevant decisions (#1041)
The CDCL-Tableaux solver uses the structure of the original formula both
to limit the literals that are propagated to the theory module and to
focus the (sub)-terms of the formula to use for instantiation.

Effectively, it only uses literals that are either asserted directly
(possibly as part of a conjunction), or that are the *first* literal
that was decided in a disjunction. For instance, if the formula [P \/ Q]
is asserted, and we decide that [P] holds, [P] will be propagated to the
theory module and used for instantiation -- but if we ever decide or
prove that [Q] holds (or does not hold), that information will be
ignored by the theory and instantiation modules.

In these conditions, deciding on [Q] is unlikely to help us make any
progress, and so this patch prevents the CDCL-Tableaux solver from
taking decisions on such irrelevant variables.