feat(UF): Store domains inside the union-find
This patch adds the ability to store domains (as used in the bit-vector,
enum and ADT theories) directly in the union-find structure, which is
responsible for updating them automatically as class representatives are
updated.

The advantages of doing this are twofold: we no longer need to manually
keep track of substitutions (avoiding the issue that some of them are
not propagated from the union-find to the relations), and having the
domains stored in a central place rather than inside each relations
module opens the path to more communication between the domains (e.g.
accessing integer domains from the bit-vector module).

Note that constraints (notably for the bit-vector theory) are *not*
integrated into the union-find and still need to be kept track of
manually.