fix(CP): Remove `ex` parameter from `Domain.intersect`
Adding an explicit `ex` parameter to the `Domain.intersect` signature
was intended to make it behave more like an update and avoid issues such
as #479. The other intent was to be a bit more efficient by avoiding
additional `add_explanation` calls.

Given that I made this design yet have now introduced the same soundness
issue twice because of it, it is now clear that this was not a good
idea.

This patch reverts that design decision and the `ex` parameter in the
`Domain.intersect` API, replacing it with explicit calls to
`add_explanation` instead.