Use the Dolmen identifiers for constructors and ADT names (#1098)
This PR fixes the way we manage incremental type declarations for the frontend Dolmen only. Before this PR we didn't manage correctly incremental type declarations. For instance the following input is accepted by the SMT-LIB standard:
```smt2
(set-logic ALL)
(push 1)
(declare-datatype t
  ((c1 (d1 Int)) (c2)))
(pop 1)
(declare-datatype t
  ((c2 (d1 Int)) (c1)))
(declare-const x t)
(declare-const y t)
(assert ((_ is c1) x))
(assert ((_ is c1) y))
(assert (distinct x y))
(check-sat)
```
(Notice that we switch the constructors c1 and c2 in our second declaration of the type t). AE answers `unknown` on next but it answers `unsat` if we remove the first declaration of t.

After this PR we got `unsat` with `--frontend dolmen` but we still got unknown with `--frontend legacy`.

The solution consists in using the unique identifiers produced by Dolmen for the constructors, destructors and names of ADTs. To reduce as much as possible the amount of modifications done in the legacy typechecker, I wrapped Dolmen's identifiers in a new module `Uid`.