Cache the slow function `update_gets_sets` (#1126)
The function `update_gets_sets` of `Arrays_rel` explores recursively
all the terms of the problem appearing in literals of `Rel.assumes`.
This function becomes a bottle neck if we are very large terms as in the
issue #1123. For instance, the input file of this issue is solved on
`next` in several minutes but instantly after this patch.

Use a local cache for Alt-Ergo-Fuzz.