fix(CLI): Do not use psmt2 format for .smt2 file
Following discussions at the Alt-Ergo Club meeting last week, it was
agreed that it was better to support SMT-LIB2 format by default for
".smt2" files and to only enable the polymorphic extensions for ".psmt2"
files.

This patch performs that change, partially reverting #948 /
42e9ef10dca19490b502db3b5f323aeb0417100e and uses the same polymorphic
variant type as Dolmen for the same purpose.

In order to accomodate #933, introduce a new "psmt2" input mode.