feat(FPA): Add ae.float{16,32,64,128} SMT-LIB builtins
This patch adds `ae.float16`, `ae.float32`, `ae.float64` and
`ae.float128` as abbreviations for `(_ ae.round 11 24)`,
`(_ ae.round 24 149)`, `(_ ae.round 53 1074)` and
`(_ ae.round 113 16494)` respectively.

These serve a similar purpose as `float32` and `float64` in the native
format: allow short names for common floating point sorts without
requiring users to remember the appropriate parameters to `ae.round`.

(Note that `ae.float16` and `ae.float128` are not available in the
native format; they are added for consistency with the `FloatXX` sorts
provided by the FloatingPoint theory in the SMT-LIB.)