Split on enum domains after asserting formulas
This PR changes the way we perform split on ADT domains.
Before this PR, after asserting formulas, we can split on ADT semantic
values of enum type, that is all its constructors have no payload.
After this patch, we can split on an ADT domain if all its constructors
are enum constructors, that is they have no payload.