Use a custom attribute to name lemma
This PR adds a new attribute ":lemma" to name lemmas instead of
using the builin attribute ":named". So the new syntax is:
```smt
(assert (! (forall ...) :lemma my_lemma))

```
Notice that we produce a custom error `Expect_simple_sexpr` if the user
gives an inappropriate s-expressions. Currently, we cannot define our
own printer for the typechecker but Guillaume will implement this soon.