Fix issue 1099 (#1102)
* Fix issue 1099

The `D_cnf` module did not use the flag `toplevel` as it was done in `Cnf`.
This flag is important because the `Expr` AST does not store quantified
type variables with binders as it does for the term variable. Instead, we
use a `prenex polymorphism` approach, which means the quantified type
variables are at the top level only.

Alt-Ergo only supports prenex polymorphism. This commit adds a failure assertion
in `D_cnf` to enforce this property.