Use Dolmen attribute to store well founded order (#1153)
* Use Dolmen attributes to store well founded order

This PR used the well designed Dolmen library (thanks Guillaume) to store a custom attribute on
constructors of ADT. This design has two advantages:
1. We do not need to store an extra global state for the total order generated in the module `Nest`.
2. We have a total order per ADT instead of a global order for all the ADT types.

To be more specific, we generate a bijection between ADT constructors and the interval [0, n[ where n is the number of constructors.

As an aside, this PR refactors the way we manage Dolmen identifiers. The new approach is safer as OCaml's typechecker can ensure some invariants for us.