chore: Update the changelog
This patch integrates changes since the 2.5.4 release into the
changelog, and imports the changelog for the 2.5.3 and 2.5.4 releases.

The development pace has picked up significantly this year, and there is
now a significant number of entries in the changelog and we might want
to consider using a toned-down version for the release notes.

Note that not all PRs are included in the changelog; in particular,
minor implementation changes or build fixes are not included. All
user-visible changes should be included.