Clean gh-pages branch (#1168)
The current `gh-pages` branch contains a copy of the `dev` branch, which is useless. 
This PR removes all the content of this branch and replace it by the content:
``` 
|-- dev 
|-- v2.5.4
|-- index.mld
```
where the directories `dev` and `v2.5.4` will content the documentation respectively of `next` and the tag `v2.5.4`. 
The `index.mld` contents links to these documentations and will be the main page of the documentation displayed by GitHub.