fix(BV): Do not lose explanations in bvmul
The implementation of bvmul from #1144 introduced a soundness bug: when
we do not know anything about the result, the explanation is dropped.

This is because the implementation was performing mixing bitlist
computation and creation of raw bitlist values. This patch fixes the
implementation by performing all computations in [Z] and only adding the
explanation at the end.