fix(CI): Ensure binary releases have proper version number
The issue boils down to GitHub not downloading tag information when
running the workflow, which means that `dune-build-info` and
`dune-subst` see a lightweight tag rather than an annotated tag and
ignore it.

In order to facilitate deployment, update the workflow to automatically
upload the artifact to the release if applicable.

Fixes #1131