feat(doc): Versioned documentation pages
This patch overhauls documentation generation to support separate
documentation pages for each Alt-ergo version on the GitHub pages
website.

We generate one version for the `next` and `main` branch, and one
version for each tag of the form `v*.*.*`. The version selector from
Sphinx is used to allow users to change the version (it appears in the
lower-right corner of the page).

We also keep a `latest` version that points to the latest tag that was
built (using a symbolic link). The default page redirects to the
`latest` version.

As part of this overhaul of the documentation workflow, the OCaml
documentation built by `odoc` is now also versioned and added to the
`API` sub-directory of the Sphinx documentation (with a placeholder to
prevent Sphinx from complaining).

Fixes #693